Add in the expiry date for the various tests to the candidate dashboard. Shift the date formatter code to the DateFormatter trait.
