+++ .gitignore
@@ -20,3 +20,4 @@ dist
 /.cache-main
 /.cache-tests
 /bin/
+.ensime*
