+++ app/views/widgets/onlineTestProgress.scala.html
@@ -40,6 +40,7 @@
                 phase2Test.getDuration
             }</strong> to finish
                 the e-tray</p>
+            <p id="eTrayExpires" class="">It expires at <strong>@{phase2Test.getExpireTime}</strong> on <strong>@{phase2Test.getExpireDate}</strong></p>
             <div id="etrayTestButton">
                 @helper.form(action = routes.CubiksTestController.startPhase2Tests(), 'novalidate -> "novalidate") {
                     @helper.CSRF.formField
@@ -195,6 +196,7 @@
                         <p id="testsTime">You have <strong>@{
                             tests.getDuration
                         }</strong> to finish the online exercises</p>
+                        <p id="testsExpires" class="">It expires at <strong>@{tests.getExpireTime}</strong> on <strong>@{tests.getExpireDate}</strong></p>
                         <div id="testsButtonContainer">
 
                             @helper.form(action = routes.CubiksTestController.startPhase1Tests(), 'novalidate -> "novalidate") {
@@ -339,6 +341,8 @@
                                     phase3Test.getDuration
                                 }</strong> to finish
                                     the video interview</p>
+                                <p id="videoExpires" class="">It expires at <strong>@{phase3Test.getExpireTime}</strong> on <strong>@{phase3Test.getExpireDate}</strong></p>
+
                                 <div id="videoTestButton">
                                     @helper.form(action = routes.LaunchpadTestController.startPhase3Tests(), 'novalidate -> "novalidate") {
                                         @helper.CSRF.formField
